Redefine a search base by not including CDbentitycontroller specific functionality. It is more general now.
